imagetypes_container: Drop oci type
This format was supported since the introduction of container images but
it was never documented. After revisiting its content and potential
purpose, it turned out that its use case would be rather theoretical and
that it can't be imported into common container services. So simply drop
it as "practically irrelevant".

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>