Create separate BOM for external initrd
The BOM (manifest file) describes all packages installed in a rootfs.
Currently, the BOM is just generated for the images rootfs, but not for
the initrd. This is fine, as long as no separate initrd is used. Tests
on isar-cip-core have shown, that ~4-10 source packages are only
referenced by the initrd rootfs, but not by the image rootfs.

We change this by generating the BOM also for the rootfs of external
initrds. Note, that this has a significant overlap with the images
rootfs, as it also contains all essential packages. As there is
currently no way to explain in a uniform BOM which components are used
where, the merging of the BOMs is left to the user.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>