speedup localepurge task by emulating less
The download and dpkg status tasks do not need to be emulated, as these
properly support running without a chroot and for a foreign architecture.

The install task itself does support running without a chroot, however
not for a foreign architecture. By that, we keep the emulation of the
install task for now.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>