linux-custom.inc: drop library dependencies from kernel headers
The upstream headers do not have dependencies to the runtime libraries.
The ${shlibs:Depends} macro will correctly generate the necessary
dependencies.

libssl is only used in extract-cert which is necessary during the kerneln build
if CONFIG_SYSTEM_TRUSTED_KEYRING is enabled.
This config element is not necessary for the module build
and is therefore removed from the kbuild packages.

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>