rootfs: use LOCALE_DEFAULT with default "C" for locale exports
Some packages, such as console-setup, fail to set a valid CHARMAP according
to the locale. For example, image-locales-extension.bbclass sets the default
locale to "en_US.UTF-8".

However, rootfs.bbclass was exporting hardcoded locale values as "C".
Because of this, when console-setup is installed in non-interactive mode
during image creation, it sets the CHARMAP based on the "C" locale, causing
a configuration mismatch.

Set LOCALE_DEFAULT to "C" by default in rootfs.bbclass so it can be overridden
by image-locales-extension.bbclass with "en_US.UTF-8", enabling proper CHARMAP
setup in console-setup.

Signed-off-by: Badrikesh Prusty <badrikesh.prusty@siemens.com>