always add architecture to binary packages in manifest
The manifest file contains the source / binary package relation and is
helpful for license clearing. However, the binary package is only
annotated with an architecture specified (:<arch>) if it otherwise
cannot be uniquely identified on the generating system. As the consumer
of that data is on a different system, this is not helpful. We now
explicitly add the architecture specifier so that downstream consumers
do not have to guess.

Despite this slightly changes the output format, we don't consider this
a breaking change, as the downstream consume anyways could not know
which packages get the architecture specifier and which not, as this
only depends on the package list of the target rootfs.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>