debianize: introduce DEBIAN_SECTION
Add support for setting the Section field in auto-generated control files.
This allows users to specify a custom section for their packages, making
it easier to categorize and identify them within the Debian archive.

By default, the section is set to "misc".

Also update the section for the isar-exclude-docs package as a demonstration.

Example:
  DEBIAN_SECTION = "ui"     # For UI-related packages
  DEBIAN_SECTION = "utils"  # For miscellaneous utilities

Signed-off-by: srinuvasan <srinuvasan.a@siemens.com>