linux-custom: add -e option to postinstall script
While inspecting the kernel package, a warning was observed related
to the postinstall script. It appears the script was not using the
`-e` option, which ensures the script stops immediately upon any error.

To fix this warning, add the -e option to the postinstall script,
similar to how other maintainer scripts have it.

Logs:
W: linux-image-cip: maintainer-script-ignores-errors [postinst]

Signed-off-by: Srinuvasan A <srinuvasan.a@siemens.com>