image: introduce IMAGE_INITRD, deprecate INITRD_IMAGE
INITRD_IMAGE could be set to the full name of an initrd image to be
found in DEPLOY_DIR_IMAGE. Downstream layers had to assume how Isar
names its image artifacts (${DISTRO}-${IMAGE}-initrd.img suffix for
initrd images) and insert the build of their initramfs recipe into
the image build pipeline.

IMAGE_INITRD is introduced to (1) have a constistent naming convention
for variables consumed by the image class (IMAGE_ prefix) (2) only
require this variable to be set when a custom initrd should be built
and used by the imager.

For instance, cip-core was doing the following:

   INITRAMFS_RECIPE ?= "cip-core-initramfs"
   INITRD_IMAGE = "${INITRAMFS_RECIPE}-${DISTRO}-${MACHINE}.initrd.img"
   do_image_wic[depends] += "${INITRAMFS_RECIPE}:do_build"

And it could now be changed to:

   INITRD_IMAGE ?= "cip-core-initramfs"

(their layer would likely keep their INITRAMFS_RECIPE variable as it
may be considered as an API towards its own downstream users).

INITRD_IMAGE is "only" deprecated; meaning that it may still be used
(but build-time warnings will be raised). If both IMAGE_INITRD and
INITRD_IMAGE are set then the latter will be ignored (a warning noting
that both were set will be emitted).

These changes were tested against (1) new CI test-cases and (2) CIP
Core 1.9.0 both with and without adaptions to its code.

Signed-off-by: Cedric Hombourger <cedric.hombourger@siemens.com>