linux-custom: fix name collision on cross profile
The cross build profile is currently used to control which packages are
built in a DISTRO_ARCH != HOST_ARCH scenario. This is fundamentally
wrong, as the cross profile is one of the standard debian build profiles
which is used to control if sbuild runs in cross mode. Further, the
build profile spec [1] forbids to conditionally generate binary packages
based on this profile. We now split this profile off into a custom one
to not interfere with the standard cross profile.

Due to the split, we now also need to set our cross profile explicitly
in the desired cases, which we now control via the cross-pkgs overrides.

[1] https://wiki.debian.org/BuildProfileSpec#Registered_profile_names

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>