rootfs: Do not recursively build unneeded packages
Using recrdeptask has the side effect that every package of which a task
appears in the dependency chain, will be built and deployed, even if just a
subset of tasks was needed to fulfil an inter-task dependency.

For coupled packages like the native/compat ones, which can share part of
their tasks, see 2ca3a7e  dpkg-source: Build source package only once,
this leads to a full build of the base package, while e.g. only the native
part needed to be produced.

Refrain from doing so by going back to using deptask which only adds the deploy
task of the direct dependencies and rely on correct cache and inter-task
settings of the used classes/recipes.
Add rdeptask handling to allow for runtime dependency settings.

This is a partial revert of 7c7628e  rootfs: recursively depend on packages.

This fixes build failures with custom "any" packages which dont support
crosscompiling (-native only), improves performance as no unneeded compiling is
being done.

Signed-off-by: Andreas Naumann <anaumann@emlix.com>