multiarch: Replace divertion of deploy task for "all" packages
Diverting the do_deploy_deb function of the base package to the native variant
had the sideeffect that the dependencies of the native variant were added to
the dependency chain even if the package was originally depended on in its non-
native form.

Fix this by not manipulating the dependencies at all, but just build the base
package for the host architecture. That way sub-dependencies are preserved for
all uses cases.

Fixes:  5a7c2f7004 ("handle DPKG_ARCH=all case for transitive deps")
Signed-off-by: Andreas Naumann <anaumann@emlix.com>