isoimage-isohybrid-isar: fix disk full if no initrd exists
If no initrd was generated during the rootfs build the wic plugin
uses the existing rootfs as initrd. The plugin uses shutil.copytree
to copy the rootfs to a new folder. As the rootfs contains `/dev/full`
this will fill the disk until it is full.

To avoid this error ignore all files in `/dev/` of the rootfs.

Tested-by:    Herbert Bernecker <herbert.bernecker@siemens.com>
Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>