initramfs-tee-ftpm-hook: Restore waiting for tpm device
It was wrong to assume that there is no need to wait for the tpm device
anymore when using in-kernel supplicant. Also in that case, there can
still be some delay in probing the eMMC storage device and, thus, also
the ftpm which depends on it. Restore our waiting loop, just updating
its comment accordingly.

Fixes: a1893507e495 ("initramfs-tee-ftpm-hook: Skip local-top hook without userland supplicant")
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>