testsuite: add parameter to emit bitbake dependency graph
When running with this avocado parameter, bitbake is called with "-g"
prior to executing the build. By that, a dependency graph of what
bitbake will execute is dumped that tremendously helps in identifying
what we test. This further helps developers to strip down tests to the
features that should be tested by locating unwanted tasks that are
pulled in via transitive dependencies.

The feature is modeled as a test function itself, so it can later be
extended to write parser-only tests or graph analysis tests.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>