improve performance by immediate evaluation of HOST_ARCH
To determine the host architecture dpkg --print-architecture is
executed. This call is currently a lazy one, hence it is evaluated at
the location where HOST_ARCH is expanded. On multiconfig this led to
almost a million process invocations, slowing down the parsing.

By making this assignment immediate, it is only evaluated once (per
recipe), reducing the parse time by around 50% on some larger
multiconfigs. Prior to the change an extra weak assignemnt was used.
However, the HOST_ARCH is not expected to change anyways, hence the
replacement should be fine.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>