Fix missing /var/lib/dpkg/available in Buster rootfs
When building an image based on Debian Buster the rootfs installation
fails on dpkg because /var/lib/dpkg/available is missing.
isar-mmdebstrap.inc tries to work around this by adding an mmdebstrap
hook. The hook uses touch /var/lib/dpkg/available, but it runs on the
host rather than inside the mmdebstrap rootfs. The proposed fix is to
use the hook parameter $1 (the mmdebstrap rootfs path) when creating
the file.

Signed-off-by: Vladimir Burkin <mxvisor@gmail.com>
Signed-off-by: Anton Mikanovich <amikan@ilbers.de>